Move discovery.ts to /query-testing/discovery.ts
